+++ webapp/src/routes/_authenticated/studies/$studyId/explore/outputs/$outputId/-components/ResultMatrixViewer/ResultFilters/DownloadVariableViewButton.tsx
