+++ webapp/src/routes/_authenticated/studies/$studyId/explore/outputs/$outputId/-components/ResultMatrixViewer/ResultFilters/VariableSelector.tsx
@@ -40,7 +40,7 @@ import {
   getVariables,
   type DataType,
   type ListType,
-} from "../../-utils";
+} from "../../../-utils";
 
 interface VariableSelectorProps {
   variablesMetadata: VariablesListDTO | null;
