+++ webapp/src/routes/_authenticated/studies/$studyId/explore/outputs/$outputId/-components/ResultMatrixViewer/ResultFilters/index.tsx
@@ -36,7 +36,7 @@ import {
   type Frequency,
   type ListType,
   type MonteCarloMode,
-} from "../../-utils";
+} from "../../../-utils";
 import ClusterSelector from "./ClusterSelector";
 import DownloadVariableViewButton from "./DownloadVariableViewButton";
 import MonteCarloModeSelector from "./MonteCarloModeSelector";
